Move files to more appropriate folders
