+++ application/ui/src/components/header/header.component.tsx
@@ -5,7 +5,7 @@
 
 import { Flex, Header as SpectrumHeader, View } from '@geti/ui';
 
-import { ProjectsListPanel } from '../features/projects-management/projects-list-panel.component';
+import { ProjectsListPanel } from '../../features/projects-management/projects-list-panel.component';
 
 export const Header = () => {
     return (
