+++ application/ui/src/features/projects-management/not-active-project/not-active-project.module.scss
