+++ application/ui/src/features/sources-sinks-configuration/sources-configuration/ip-camera/ip-camera.component.tsx
